refactor: 파일 이름 변경 및 코드 정리
