+++ frontend/src/pages/AnalysisResult.css
@@ -1,0 +1,63 @@
+.result-container {
+  max-width: 900px;
+  margin: 0 auto;
+  padding: 30px;
+}
+
+.selected-info {
+  background-color: #f1f5f9;
+  padding: 15px;
+  border-radius: 8px;
+  margin-bottom: 20px;
+}
+
+.sections-container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.section-card {
+  background-color: #ffffff;
+  border: 1px solid #e2e8f0;
+  border-radius: 10px;
+  padding: 20px;
+  box-shadow: 0 4px 8px rgba(0,0,0,0.05);
+  transition: transform 0.2s;
+}
+
+.section-card:hover {
+  transform: translateY(-4px);
+}
+
+.section-card h3 {
+  margin-bottom: 10px;
+  color: #0f172a;
+}
+
+.section-card p {
+  margin: 0;
+  line-height: 1.6;
+  color: #334155;
+}
+.header-bar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 20px;
+}
+
+.create-roadmap-btn {
+  padding: 8px 16px;
+  background-color: #2563eb;
+  color: white;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  font-weight: bold;
+  transition: background-color 0.2s;
+}
+
+.create-roadmap-btn:hover {
+  background-color: #1d4ed8;
+}
